rpmsg: smd: Represent smd edges as devices
By representing each edge as its own smd edge the channels are no longer
tied to being parented by the same smd device and as such can live as
children of e.g. remoteproc instances.

It also makes it possible to have use device matching operations based
on channel name to find a particular smd device, as channel names are
unique per edge, but not in the bigger system.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>