remoteproc: Qualcomm SSR proof of concept
This POC implements the necessary mechanism for tying smd edges to a
remoteproc and control the edge's life cycle based on the remoteproc's.

The result is smd devices being registered when the remoteproc boots,
they are properly reset on a crash and they are closed when the
remoteproc is shut down.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>