venus: Turn venus into a child of venus-rproc
By making venus a child of the venus rproc it will follow the running
state of the venus processor rather than requesting it to be booted
following the venus driver.

This makes it possible to utilize the remoteproc system for controlling
when the venus should be initialized, based on e.g. availability for
firmware and makes the venus driver follow the state of the venus
processor during crash recovery.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>