remoteproc: add IPA notification to q6v5 driver
Set up a subdev in the q6v5 modem remoteproc driver that generates
event notifications for the IPA driver to use in implementing
subsystem restart (SSR).

This sub-device is only used by the modem subsystem (MSS) driver, so
some of the code implementing it is found in "qcom_q6v6_mss.c" (as
opposed to "ipa_common.c").

A pair of new functions provides a way for the IPA driver to register
a notification callback function that will be called whenever modem
events (about to boot, running, about to shut down, etc.) occur.
The callback is supplied a void pointer value (provided by the IPA
driver at registration time) and an event type.

One event, MODEM_REMOVING, is signaled whenever the q6v5 driver is
about to unregister the notification subdevice.  It requests that the
IPA driver de-register this callback.

Signed-off-by: Alex Elder <elder@linaro.org>