soc: qcom: ipa: IPA interface to GSI
This patch provides interface functions supplied by the IPA layer
that are called from the GSI layer.  One function is called when a
GSI transaction has completed.  The others allow the GSI layer to
inform the IPA layer when the hardware has been told it has new TREs
to execute, and when the hardware has indicated transactions have
completed.

Signed-off-by: Alex Elder <elder@linaro.org>
Link: https://lore.kernel.org/r/20200228224204.17746-9-elder@linaro.org
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>