phy: qcom-qmp: Register typec mux and orientation switch
The QMP PHY handles muxing of USB vs DisplayPort, as well as orientation
switching of the SuperSpeed lanes. So register typec handlers for the
two types.

The TypeC mux allows switching between four lanes of DisplayPort and a
mixed USB+DP combination. This makes it possible to reach resolutions
that requires 4 lanes.

The TypeC switch allows switching the SuperSpeed pins and have been
tested with both 2 and 4 lane DisplayPort.

It's possible that in the USB mode the DP_MODE should be disabled, but
this is left untouched.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>