Revert "soc: qcom: rpmhpd: Make power_on actually enable the domain"
This reverts commit e3e56c050ab6e3f1bd811f0787f50709017543e4.

The behaviour prior to the introduction of 'e3e56c050ab6 ("soc: qcom:
rpmhpd: Make power_on actually enable the domain")' was to simply ignore
enable requests when the performance state was 0. While this isn't
correct, it's now possible that enable will vote for a lower performance
state than what is required by the enabled hardware, in particular
during boot. So revert this temporarily.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>