arm64: dts: qcom: sa8295-adp: Add PMICs and related GPIO nodes
The SA8295p ADP has four PM8540 PMICs.

The PM8540 is, based on it's version information, compatible with the
PM8150, but rather than creating four copies of pm8150.dtsi the four
PMICs are defined in the board file for now.

Also define the GPIO block in each one, to allow driving the USB enable
pins spread over three of the four PMICs.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>